+++ src/RmqCli/Commands/MessageRetrieval/BaseMessageRetrievalService.cs
@@ -3,9 +3,9 @@ using System.Threading.Channels;
 using Microsoft.Extensions.Logging;
 using RabbitMQ.Client;
 using RmqCli.Core.Models;
-using RmqCli.Core.Services;
-using RmqCli.Infrastructure.Output;
+using RmqCli.Infrastructure.RabbitMq;
 using RmqCli.Shared;
+using RmqCli.Shared.Output;
 
 namespace RmqCli.Commands.MessageRetrieval;
