+++ src/RmqCli/Commands/MessageRetrieval/Strategies/PollingStrategy.cs
@@ -3,8 +3,9 @@ using System.Threading.Channels;
 using Microsoft.Extensions.Logging;
 using RabbitMQ.Client;
 using RmqCli.Core.Models;
-using RmqCli.Infrastructure.Output.Formatters;
 using RmqCli.Shared;
+using RmqCli.Shared.Output;
+using RmqCli.Shared.Output.Formatters;
 
 namespace RmqCli.Commands.MessageRetrieval.Strategies;
