+++ src/RmqCli/Commands/Peek/PeekCommandHandler.cs
@@ -1,6 +1,8 @@
 using System.CommandLine;
 using RabbitMQ.Client.Exceptions;
 using RmqCli.Shared;
+using RmqCli.Shared.Factories;
+using RmqCli.Shared.Output;
 
 namespace RmqCli.Commands.Peek;
