+++ src/RmqCli/Commands/Publish/PublishResponseFactory.cs
@@ -1,5 +1,6 @@
 using RmqCli.Core.Models;
 using RmqCli.Shared;
+using RmqCli.Shared.Output;
 
 namespace RmqCli.Commands.Publish;
