+++ src/RmqCli/Commands/RootCommandHandler.cs
@@ -3,6 +3,7 @@ using RmqCli.Commands.Config;
 using RmqCli.Commands.Consume;
 using RmqCli.Commands.Peek;
 using RmqCli.Commands.Publish;
+using RmqCli.Shared.Factories;
 
 namespace RmqCli.Commands;
