+++ src/RmqCli/DependencyInjection/ServiceCollectionExtensions.cs
@@ -7,11 +7,11 @@ using RmqCli.Commands.MessageRetrieval;
 using RmqCli.Commands.MessageRetrieval.Strategies;
 using RmqCli.Commands.Peek;
 using RmqCli.Commands.Publish;
-using RmqCli.Core.Services;
 using RmqCli.Infrastructure.Configuration;
 using RmqCli.Infrastructure.Configuration.Models;
-using RmqCli.Infrastructure.Output;
+using RmqCli.Infrastructure.RabbitMq;
 using RmqCli.Shared;
+using RmqCli.Shared.Output;
 
 namespace RmqCli.DependencyInjection;
