+++ src/RmqCli/Shared/Output/FileOutput.cs
@@ -4,10 +4,9 @@ using System.Threading.Channels;
 using Microsoft.Extensions.Logging;
 using RmqCli.Core.Models;
 using RmqCli.Infrastructure.Configuration.Models;
-using RmqCli.Infrastructure.Output.Formatters;
-using RmqCli.Shared;
+using RmqCli.Shared.Output.Formatters;
 
-namespace RmqCli.Infrastructure.Output.File;
+namespace RmqCli.Shared.Output;
 
 /// <summary>
 /// Writes consumed messages to file(s).
