+++ src/RmqCli/Shared/Output/Formatters/TableMessageFormatter.cs
@@ -1,11 +1,10 @@
 using System.Text;
 using RabbitMQ.Client;
 using RmqCli.Core.Models;
-using RmqCli.Shared;
 using Spectre.Console;
 using Spectre.Console.Rendering;
 
-namespace RmqCli.Infrastructure.Output.Formatters;
+namespace RmqCli.Shared.Output.Formatters;
 
 /// <summary>
 /// Formats RabbitMQ messages as pretty-printed tables using Spectre.Console.
