+++ test/RmqCli.Integration.Tests/Infrastructure/Output/FileOutputTests.cs
@@ -4,10 +4,9 @@ using Microsoft.Extensions.Logging.Abstractions;
 using RabbitMQ.Client;
 using RmqCli.Core.Models;
 using RmqCli.Infrastructure.Configuration.Models;
-using RmqCli.Infrastructure.Output;
-using RmqCli.Infrastructure.Output.File;
-using RmqCli.Infrastructure.Output.Formatters;
 using RmqCli.Shared;
+using RmqCli.Shared.Output;
+using RmqCli.Shared.Output.Formatters;
 
 namespace RmqCli.Integration.Tests.Infrastructure.Output;
