+++ test/RmqCli.Integration.Tests/Infrastructure/Output/MessageOutputFactoryTests.cs
@@ -1,10 +1,8 @@
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Logging.Abstractions;
 using RmqCli.Infrastructure.Configuration.Models;
-using RmqCli.Infrastructure.Output;
-using RmqCli.Infrastructure.Output.Console;
-using RmqCli.Infrastructure.Output.File;
 using RmqCli.Shared;
+using RmqCli.Shared.Output;
 
 namespace RmqCli.Integration.Tests.Infrastructure.Output;
