+++ test/RmqCli.Unit.Tests/Infrastructure/Output/Formatters/TextMessageFormatterTests.cs
@@ -2,8 +2,9 @@ using System.Globalization;
 using System.Text;
 using RabbitMQ.Client;
 using RmqCli.Core.Models;
-using RmqCli.Infrastructure.Output.Formatters;
 using RmqCli.Shared;
+using RmqCli.Shared.Output;
+using RmqCli.Shared.Output.Formatters;
 using RmqCli.Unit.Tests.Helpers;
 
 namespace RmqCli.Unit.Tests.Infrastructure.Output.Formatters;
